"Fixed fragment-tolerance. Accidently misreprested as ppm instead of dalton"

+++ src/app/tide/spectrum_preprocess.h
@@ -151,7 +151,7 @@ class ObservedPeakSet {
                                    double precursorMass,
                                    int nAA,
                                    const vector<double> aaMass,
-                                   int fragTol,int granularityScale,
+                                   double fragTol,int granularityScale,
                                    vector< vector<double> >& residueEvidenceMatrix);
   //END -- added by Andy Lin
